[Ref] intial testing on case tokens, make knownTokens optional
There is a perception in the token code that passing around knownTokens is more performant

This may or may not be true - but since the goal is to migrate the code to the
token processor making the code simpler is
a higher priority as any performance change will change again.

This makes knownTokens optional in the replaceCaseTokens and adds test cover.
It is a step towards simplification in the PdfLetter code
which could be simplified if not needing to pass in knowntokens here