REF Copy preProcessFromAddress back into the pdf function
This is really hard to unravel without splitting it between classes -
mostly because it does stuff to do with setting properties on the form
that are not even relevant to all forms. This is a divide & conquer
case to cleaning up the function from
the deprecated class

Note that I removed the bounce on no-email. It felt like we would still want
to be able to download pdfs if the logged in user had no email. The bounce param seems
to have been added as an after thought - presumably to block
the bounce in a flow where they realised it made no sense