[REF] dev/core#2790 Move getFileName functionality to the trait
This moves the recently added functionality to standardise the file name to PDFTrait

For code not using the trait (naughty extensions that are using core code in unsupported ways)
the functionality will be the same as prior versions.

(note extensions should just copy & paste the code they want from core into their extensions
rather than call core functions - it's not 'DRY' if it's not supported).