dev/core#2832 Add test for legacy membership tokens, add support for preferred tokens
There is only one place that calls this code. With cover of the legacy tokens
I can ensure it is the same as the processor tokens. I've fixed it
to support the preferred unambiguous tokens because that means that
we don't have to deal with reports of them not working while calls
to it are phased out of extensions. Working through that
also allowed me to remove some code now