[REF] simplify member_tokens
Now we have good test cover we can switch to using the parent.

The one thorn in our side is membership.fee.
Dealing with the field, and hence opening up 'listening'
is out of scope for this pr

We can (and should) switch fee to membership_type_id.minimum_fee and add
generic support. However, we also have a formatting issue.
The field is formatted 'number only' 'just cos'.

If we map that field without figuring out formatting
it will be locked in as an anomaly.