Civi\Test::example() - Make it a bit easier for the typical case
This changes the signature on a new helper method. This method is not widely used, and each reference is updated here.

Before: `Civi\Test::example($name)` returns the *metadata* for the example.

After: `Civi\Test::example($name)` returns the *data* for the example.

Comment: It's more convenient to stitch together examples from the data.  Of
course, metadata may also be useful -- it's still available through
`Civi\Test::examples()->getFoo(...)` (with a few different `getFoo()`
methods).