Convert event badges to use token processor
This adds the token processor class for participants and
switches the one place where we currently resolve participant
tokens to use it.

In the process I upgrade the badge tokens with a syntax change and
drop a few more that are actually
a) weird / seemingly accidentally exposed and
b) not really part of the participant entity

ie event template title & default role id

With participant tokens we have been fortunate not
to let a non-standard implementation go out
so it's only the event badges which are
a bit obscure / less used / less likely to be
blindly used. I have added token replacement
but I think it will be all very edge case