[REF] Extract getFromValues, add test
This extracts the code used to get the from values, moves the default setting
to the default and adds a test.

The goal is actually to add a test on the submit function
but it's proving to be challenging so this at least gets the
default setting tested.

Note that this code used to be on a static class, not a trait, and
some stuff is in weird places - in particular stuff that
should be in setDefaults and postProcess is happening on
preProcess. However, most
of the classes that override this
trait do not do much more than call the trait so that helps at
least