Fix bug where log hook is ignored on custom field create
It is possible to specify that various tables should not
result in a log table via the `alterLogTables` hook.

This is useful for excluding tables of low information
value or tables like summary fields that are calculated
values. However, when altering custom fields this
hook's output was being ignored as it was going through
the specific path of
which did not have awareness of this hook.

This code path is the primary use for this function.

It's also called via createMissingLogTables - but in
that case the calculation is already done. It is not
elsewhere in git universe and the return value
is never used, so I removed it