dev/core#2814 Add support for preferred syntax for contact tokens
This maintains support for all currently advertised tokens. But it
switches the advertisement to our preferred style (matching the db field,
specifically denoting if the label is desired.) Tests for both styles are added.

This is a step towards switching to apiv4. An important difference with v4 is
that we can distinquish between 'not fetched' and 'empty' - saving us
from extra queries when it is not clear - which will in turn
allow us to switch greeting
processing over & still support the performance tweaks via
that method