ensure recur options are present on backend cc contribution form.
We seem to be adding an extra check to see if the given
payment processor has any credit card fields before including the
recur option on the backend contribution page. But some processors,
like Stripe, insert the fields via javascript. It should be enough
to just test for the existence of a payment processor that supports
recur to run the block.

Also, no other code is checking the return value of buildPaymentForm
so removing it.