Return the sendEmail function to it's owner
This sendEmail function is only called from one place in core and it is not 'generally useful' having
an awful parameter set. This PR moves it back to the class that 'owns' it - which will
allow us to undo all the work of building up that parameter set
and make it possible to support tokens for other entities than those already mangled in.

I would normally add a noisy deprecation notice once a function becomes unused in
core but since that has been done to the pdf task this release I've left this
deprecation a bit quieter for now.

Note that I cleaned up the tokens handled here before deprecating so we
could get rid of those calls fully

Under the OO structure it becomes easier to add the missing token options
- membership & participant - but the business of 'one email per person, &
just grab the tokens from the last entity' is messing with
my head a bit. That's the next bit....