Afform - Set default values from URL with or without field prefix
For input forms, either of these will work:
  #?first_name=Bob
  #?Individual1.first_name=Bob

For search forms, the display name is used as the field prefix (and range values are separated with a dash). Either of these will work:
  #?total_amount=500-1000
  #?MySearchDisplay.total_amount=500-1000

Note that this assumes the main search entity is Contribution. For joins, the name of the join is required e.g. #?Contribution_Contact.first_name=Bob