Move getCustomFieldTokens to the tokens class
This deprecates (quietly) another function on CRM_Utils_Token. It's a pretty small function -
the sort you'd just copy & paste if you wanted to use outside of core but
I didn't add noise. The only core place that calls this is the tokenTrait
- which is 'used' by activity tokens - but only for the handling of _N_
which I think was actually not supposed to have been included in the merge.

Anyway - I'm leaving that question for later because I'm not sure if the token
trait is used outside of core or how our final interface will look