Remove now-obsolete getPseudoTokens
This is now only called from legacy code. The buildContributionArray path
is still a little used :-( but is tested
from CRM_Contribute_Form_Task_PDFLetterCommonTest
and hard-coding the array now it is settled is appropriate
to that code path's life expectancy

This code is no longer used in the Tokens classes and the functions that used it
from outside of there are now 'settled into their final state & waiting to die'
so a hard-coded list will see out their lives. They are marked now as deprecated