dev/core#2907 fix for gdpr clobering contact tokens
This is a fix for contact tokens not rendering when gdpr tokens is installed.

Gdpr declares some addtional tokens with the namespace 'contact'. I am inclined to
agree this counts as 'naughty' and so while I have fixed the clobbering by
not processing contact-metadata tokens I have not fixed the enotices.

If we think it is non-naughty I can kill those - but I've already loosened
some strictness that I think had benefits in order to facilitate this