(REF) Tokens - Populate "Event"/"Participant" without direct access to actionSearchResult
Overview
--------

When sending a scheduled remidner for a `civicrm_participant` record, it supports
tokens for both `{participant.*}` (`$context['participantId']`) and `{event.*}`
(`$context['eventId']`). This creates a special requirement to load the `eventId`
in addition to the `participantId`.

This patch changes the way in which `eventId` is loaded.

Before
------

In `ParticipantToken`, the token-evaluation-step for `{participant.*}`
tokens has a side-effect of copying `$context['eventId']` based on reading
`actionSearchResult` (although this appears indirect).

After
-----

The action-search returns `tokenContext_eventId`, which automatically maps
to `$context['eventId']`.

This means that the `eventId` will be set before evaluation begins.