Remove unnecessary code
This function is called from 2 places

Contact::create
Contact::updateGreetingsOnTokenFieldChange

In both cases the contact has already been saved (to the extent that is relevant
for the operation). The latter operation will not change
the selected options - it will only potentially change tokens within it. In the
former case the nulling will already have taken place if needed ie if there
is no id entity for the relevant greeing
ensureGreetingParamsAreSet will already have set it to 'null'
before the save - leading to it already being updated to NULL