[NFC] [REF] Test class cleanup
The driver for this was to remove uses of the DAO->createTestObject methods which were creating
individuals with data invalid for them. However, this is a full 'margin cleanup' and
standardises the test.

I think the whole createTestObject approach is best viewed as an early attempt to get
lots of entities testable quickly but it's overwraught for creating contacts
(and generally a bit flakey)