Remove isset from tpl file
Per https://github.com/civicrm/civicrm-core/pull/21935 it turns out isset will be
a blocker to ever doing escape on output so switching a couple of places to
empty. In the last case empty doesn't cut it so I added smarty:nodefaults
so that it would not be escaped.

isset doesn't work on the out put from a function