ManagedEntity - Add update mode 'unmodified' and fix cleanup mode 'unused' for APIv4
Update mode 'unmodified' will only update a record if it has not been locally edited.
This new setting works only for entities opted-in to the APIv4 ManagedEntity trait, and will
emit a warning and fall back on 'always' for others.

Cleanup mode 'unmodified' now works for APIv4 managed entities, and they are cleaned up
in reverse order to ensure references are deleted before their parents.