EventChecker - Fix recent regression
Overview
--------

v5.43 introduced `Civi/Test/EventChecker.php` as part of the execution of all unit-tests.

`civix` generates unit-tests from various templates. As part of the `civix` update-testing, it makes sundry tests with sundry phpunit.

Before
------

In 5.43.0, when running `civix`'s` `tests/make-example.sh`, it fails on one of the tests when using an older version of phpunit.

After
-----

When running `civix`'s` `tests/make-example.sh`, it passes with all tests.

Comment
-------

The type-hint is too strong -- the actual type-name depends on the version of phpunit. It is, of course, good
to have type-hints, and this preserves the softer `@param` hint.