dev/core#2968 - Fix invalid assets url paths
On instances that use symlinks in the configuration;
assets urls are not generated properly due to the
fact that real paths are stored in the $civicrm_root variable
while the $cmsPath might use the symlink path.
This needs to be normalized before we can use str_replace,
or we'll end up with the system path in the generated urls.