APIv4 - Delete unused OptionList trait
This trait had been added for the sake of SearchKit to determine which entities should be
searchable and which were merely supplying option lists for other entities.

However, this logic broke down quickly as some option lists *should* be searchable,
so that path was abandoned in favor of the `@searchable` annotation.

The trait now does nothing and is just taking up space and adding unnecessary complexity.