(REF) CRM/Upgrade - Remove unused entrypoint `verifyPreDBstate()`
Overview
--------

In olden times, upgraders sometimes implemented the method `verifyPreDBstate()`. However, this appears to be unused.
Additionally, it's not clear (from the documentation) how you would use it (if you wanted to).

So...  simplify the code....

Before
------

Function exists. But it does nothing. Because no subclasses implement it.

After
-----

Function does not exist. It still does nothing.