Remove handling for always truthy var being false
In a previous unshared life the code in this function supported contact being null

However, in it's unshared version contactID is always an integer (in the merge context). This strips
out the now-irrelevant code.

Further cleanup to follow ... it's not clear this function is doing
anything of value at all but some unwravelling required