[REF] Duplicate & unshare processFormContribution
The only way I've managed to break up these big toxic functions is divide & conquer. While it seems
counter-intuitive to not doing things in more than one place in practice much of the
code is form-relevant and the the code before, during and after these functions
does a whole lot of extra work to share stuff that they don't really have in common.

This dates back pre-api when the only copy of the business logic often was on the forms....