[REF] Simplify getContributionStatuses
The effect of the altered calls to getContributionStatus is to get a small set of options returned.

In the cases of participant & membership forms this is pending or completed. For
the ufGroup option it seems cancelled & failed are also data entry options (this method
should still cope if they don't exist).

The existing function is really confusing - but all that confusion is designed for one
form - back office contribution form - and augmented by overloading by other functions.

This splits it out.

Could sites be using 'weird and wonderful other statuses'? Our system really is set up
around supporting specific statuses and doing financial entries for those so I think maybe
we are right to not try to pander to any that could exist