"Open in new tab" on .crm-summary-link icons should open contact page, not summary content.
Whilst the contact image tooltip is set to preventDefault, it is still possible to copy the URL or open in new tab.
This change ensures that the correct URL is set in these situations.
Backwards compatiability is retained for third parties using the crm-tooltip class without the new data attribute.