Afform - Add field types + pseudoconstant metadata to AfformSubmission entity
The html type info is needed for the fields to be usable in the Afform GUI.

The reason for adding the pseudoconstant is to facilitate automatically fetching
the title from the afAdminFormSubmissionList form.

This could be inefficient for sites with thousands of forms,
but I think that's sufficiently mitigated by the fact that
this pseudoconstant won'be be used for much;
it won't generally be loaded onto the client-side.