APIv4 - Don't suppress 'required' attribute in getFields
Before: APIv4 getFields would always set 'required' to FALSE unless the action was 'create'
After: Attribute is correctly passed through regardless of action

The original thinking behind this suppression was that only 'create' requires values,
and 'update' doesn't need anything but an ID.
However, that logic is enforced elsewhere, and doesn't need to be baked into getFields.
And sometimes it's useful to know if a field is required, e.g. in the SearchKit
in-place edit UI, it should be conveyed to the user if they are allowed to submit
an empty value for a given field.