SearchKit - Enable joins for custom fields and option groups
There are a few FKs in the database that also have a pseudoconstant.
Arguably there shouldn't be, it should be one or the other. But there they are.
Previously SearchKit would ignore those FKs, but they are needed for e.g.
searching for custom groups and displaying the aggregate count of fields in that group.