APIv4 - Add `nullable` property to getFields; improve SearchKit editable UX
Unlike the 'required' field property, which only determines if the API requires a value to Create,
the 'nullable' property tells a UI whether a field is allowed to be set to NULL in Create OR Update.

SearchKit uses this property during in-place edit and bulk edit operations to determine whether a
field can be left blank.