Stop calling hook twice
Per the code comments this hook is called here but also in add. This is all about the crazy
history of this function - which used to be tied to the relationship form.

I can't think of any valid use for calling the hook twice - but I can think of reasons it
would be bad...