Enotice fixes
I grepped for 'editor' being assigned & didn't find it. I think changed the editor to 'TextArea'
and didn't see the text inside the if so I concluded it is broken. Removing rather than
fixing cos no-one noticed it was broken so can't have been important (and probably the
wrong place for it)