Remove handling for impossible array
CRM_Financial_BAO_FinancialType::buildPermissionedClause is only called from one place
outside of tests. That place passes in this->_whereClause which is the result of calling
this->whereClause() a few lines earlier. As that function always returns a string
we know that it cannot be an array & can remove the handling