Remove unreachable, nonsensical code
This block of code appears to be unreachable, because the contribution page
wizard is never frozen in "view" mode.
If it were reachable, it would do nothing sensible, as the button would link
to a completely unrelated page (configure custom data) and the link would be
broken because it is incorrectly bypassing `CRM_Utils_System::url()`.