Remove isset from online contribution receipt
This removes the isset checks that re failing around
totalTaxAmount (solved the same way as offline
https://github.com/civicrm/civicrm-core/pull/22560/files#diff-fd5668d5492e5f0ec55b7f8a10eccfa4ea9de249e4cf2383a2e0d36dbd92ebe0R134
)

It also removes the isset around contribution.amount_level (switching
to a token) and the enotice on isQuickConfig - this
switches to loading that information & also using a more meaningful variable.
Note that line items should always be assigned so that check is silly.