ExtensionUpgrades - Skip trying to upgrade missing dependencies
Before: Missing dependencies would be added and then the extension upgrader would try and fail to upgrade them
After: They are skipped, as missing extension dependencies are handled elsewhere

This allows the extension upgrade to proceed without error, even if there are missing dependencies.
The user will be prompted to install the missing dependencies afterward.