APIv4 - Deprecate and stop using PreSaveSubscriber
The PreSaveSubscriber allowed APIs to tweak inputs before saving a record,
but it was unreliable because:
- It would run for Create and Update but not Save actions
- It would run before pseudoconstant suffixes had been resolved
- It would run before any formatting or internal processing had taken place

Meanwhile, a new pattern of overriding the create/save/update actions and sharing a trait
has become more common and while a bit more cumbersome to implement, it does not share
those limitations.

This removes all uses of PreSaveSubscriber, refactors those bits into Save traits, and
adds a deprecation warning in case any extension in the Universe happens to have implemented it.