APIv4 - Add entity metadata about class args
The CustomValue API is a virtual API, where multiple entities all get routed to the
same class if they share the prefix "Custom_", and pass a class arg to the
php factory functions e.g. `CustomValue::get('MyCustomGroup')`.
Instead of hard-coding this idea into the  now it's part of the entity
metadata so that other APIs, notaby ECK, can use a similar pattern.