(NFC) LoggingDetailTest - Improve reliability of test
Before
-----------------

If you run this test repeatedly, it may alternately pass and fail.  The
percentage of passes-vs-fails can vary widely (depending on the computer
where you execute).

To test this, I went to a deployment (on `test-3`) which had a fairly
high probability of failure (*failure in roughly 3/4 invocations*).

After
-----------------

On the same system, it passes consistently (*no failures after 15 invocations*).