[REF] Deprecate BAO_Contact::retrieve
This removes 2 calls to BAO_Contact::retrieve and marks it as deprecated.

It is a surprisingly complex function and all the calls to it only use parts of what it
does - in general just having the code 'do the thing' is better.

In this case it turns out there is no need for the contact objects. There is
also some pretty funky handling for the location keys in the next section
so making it clearer what is in them should make it possible to simplify that....