Re-fix enotice on templateSelected
The empty check on templateSelected was added to stop tests failing on a smarty notice.

However, this still causes a smarty notice (and test fails) in grumpy smarty mode.

Hence this re-fixes in by moving the check to only be done for the from
(the sms Upload form) that assigns it