Finish conversion of static function to non-static
The processSecondaryContribution function is no longer static but the class still passes itself
to the function as the form parameter. This updates calls to form to calls to this.

Also for clarity the function is made private

I did do a quick universe search to confirm no calls to it - even thought they
would have been clearly unsupported