Start to fix misnamed FinancialTrxn bao
The FinancialTrxn BAO is misnamed in that it does not match the DAO file structure

Pradeep started to try to address this but it stalled on some test
issue & went stale

This is a baby-step on that which puts the right file there & moves
the create function, while not touching the rest.

The function doesn't refer to 'self' so should be all good.

I obsevered some of the functions are NOT used....