[Import] [Ref] Iterate through the mapping of fields rather than a 'count'
The mapper holds an array of mappings for each of the active fields.
The active field count holds a count of this array - switching to
a foreach loop means we can stop calculating extra stuff.

In addition the mapper actually has everything we need to do the rest of
the stuff in this function - which is not in this PR but
will involve a lot of code simplification