[Import] [Ref] Iterate through the mapping of fields rather than a 'count'
The mapper holds an array of mappings for each of the active fields.
The active field count holds a count of this array - switching to
a foreach loop means we can stop calculating extra stuff.

This PR doesn't start ripping that out - for legibility reasons - but it
will take out a whole lotta code....