(E2E) `testErrorChrome` - Update to work with current wp-demo build
This test was added during 5.50.alpha (#23257, #22805).  It's purpose is to
assert that the page-chrome/site-wide-template is present.  It appears that
the markup from WordPress has changed slightly.  (It may be that I
originally tested against an older version of WP.) In any event, this
relaxes the test so that it also passes on a newer WP build.