FiveFortyNine - Don't munge soft credit 'limit-to'
I'm not sure what it even means to limit a scheduled-reminder to a "soft_credit_type".

But it's a thing - I can use the 5.48 GUI to create it. Here's how it looks in the DB (in 5.48):

+----+------------------------------------------+------------------+----------+------------------+-------------------+----------+
| id | title                                    | recipient        | limit_to | recipient_manual | recipient_listing | group_id |
+----+------------------------------------------+------------------+----------+------------------+-------------------+----------+
| 11 | Limit To - Soft Credit                   | soft_credit_type |        1 | NULL             | in_memory_of      |     NULL |
+----+------------------------------------------+------------------+----------+------------------+-------------------+----------+

The `changeBooleanColumnLimitTo()` was coercing down to `limit_to=null` because it lacked `recipient_manual` and `group_id`.