Handle error when creating related contact
This addresses the e-notice identified here
https://github.com/civicrm/civicrm-core/pull/23643#discussion_r887406505

I believe the bug report in that thread was a variant of the e-notice. Ie the import()
function was failing to catch an exception & it was bubbling up to run()

While there is no support currently for calling import()
outside of run() - the intent is to replace run() with something
more sane - which WILL be callable from the outside (likely
via api).

The right behaviour for import() is to catch all errors