AngularJS - Remove UI-Utils library
The [UI-Utils Libary](https://github.com/angular-ui/ui-utils/tree/v0.1.1) was introduced in 2014,
but never used for much of anything. It remains very underutilized, and is the source of a major bug
(see https://lab.civicrm.org/dev/core/-/issues/2688).
Since it's abandonware, it seems best to stop using it rather than try to patch it.

Fixes dev/core#2688